requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
pandas==2.0.3
requests==2.32.4
flask==2.3.3
flask-socketio==5.3.6
